Added metadata generation
