+++ src/app/layout.tsx
@@ -1,6 +1,5 @@
 import { apiPlugin, storyblokInit } from '@storyblok/react';
 import StoryblokBridgeLoader from '@storyblok/react/bridge-loader';
-import { Metadata, Viewport } from 'next';
 import { headers } from 'next/headers';
 import Script from 'next/script';
 import PlausibleProvider from 'next-plausible';
@@ -14,70 +13,9 @@ import PosthogPageView from '@/components/analytics/PosthogPageView';
 import BlogNewsletterForm from '@/components/blog/BlogNewsletterForm';
 import Footer from '@/components/layout/Footer';
 import Header from '@/components/layout/Header';
-import { siteConfig } from '@/constants/config';
-import { isProd } from '@/constants/env';
 import PosthogProvider from '@/providers/posthog';
-import { openGraph } from '@/utils/og';
 
 import components from '../../storyblok';
-
-const url = siteConfig.url;
-const title = siteConfig.title;
-const description = siteConfig.description;
-const ogImage = openGraph({
-  title: siteConfig.tagline,
-  description,
-});
-
-export const metadata: Metadata = {
-  title: {
-    default: `${title}: ${siteConfig.tagline}`,
-    template: `%s | ${title}`,
-  },
-  description,
-  robots: isProd
-    ? { index: true, follow: true }
-    : { index: false, follow: false },
-  icons: {
-    icon: '/favicon.ico',
-    shortcut: '/favicon-16x16.png',
-    apple: '/apple-touch-icon.png',
-  },
-  manifest: '/site.webmanifest',
-  alternates: {
-    types: {
-      'application/rss+xml': [
-        { url: '/blog/rss.xml', title: 'Fix Security blog RSS feed' },
-      ],
-      'application/atom+xml': [
-        { url: '/blog/atom.xml', title: 'Fix Security blog Atom feed' },
-      ],
-      'application/json': [
-        { url: '/blog/feed.json', title: 'Fix Security blog JSON feed' },
-      ],
-    },
-  },
-  openGraph: {
-    url,
-    title,
-    description,
-    siteName: title,
-    images: [ogImage],
-    type: 'website',
-    locale: 'en_US',
-  },
-  twitter: {
-    card: 'summary_large_image',
-    title,
-    description,
-    images: [ogImage],
-  },
-};
-
-export const viewport: Viewport = {
-  themeColor: '#3d58d3',
-  colorScheme: 'only light',
-};
 
 storyblokInit({
   accessToken: process.env.STORYBLOK_OAUTH_TOKEN,
@@ -144,7 +82,6 @@ export default function RootLayout({
             </Suspense>
           </main>
           <Footer />
-          {/*<CookieConsent/>*/}
           <Suspense>
             <PosthogPageView />
           </Suspense>
