--- src/app/page.js
@@ -1,6 +1,0 @@
-import PageComponent from '../components/PageComponent';
-
-export default async function Page() {
-  const slug = 'home';
-  return <PageComponent slug={slug} />;
-}
