--- src/lib/storyblok.js
@@ -1,13 +1,0 @@
-import { getStoryblokApi } from '@storyblok/react';
-
-export const fetchStory = async (slug) => {
-  const storyblokApi = getStoryblokApi();
-  try {
-    let { data } = await storyblokApi.get(`cdn/stories/${slug}`, {
-      version: 'draft',
-    });
-    return data?.story || null;
-  } catch (error) {
-    return null;
-  }
-};
